% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{abs_ppm_diff}
\alias{abs_ppm_diff}
\title{Calculates differences in absolute parts per million}
\usage{
abs_ppm_diff(x, y)
}
\arguments{
\item{x, y}{vector numeric}
}
\value{
vector double
}
\description{
Calculates differences in absolute parts per million
}
\keyword{internal}
