% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{bdc_dsets}
\alias{bdc_dsets}
\title{Title}
\usage{
bdc_dsets()
}
\value{
vector character: Names of valid bdc datasets
}
\description{
Title
}
\seealso{
Other bdc functions: \code{\link{bdc_filter_mz}},
  \code{\link{bdc_filter_rt}},
  \code{\link{bdc_mzid_match}},
  \code{\link{bdc_pick_rt_interval}}
}
