% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{bdc_mzid_match}
\alias{bdc_mzid_match}
\title{Title}
\usage{
bdc_mzid_match(x, table, ppm_thresh = 10)
}
\arguments{
\item{x, table}{list: Each element should be an mzid object.}

\item{ppm_thresh}{scalar numeric: ppm threshold to filter mz differences}
}
\description{
Title
}
\seealso{
Other bdc functions: \code{\link{bdc_dsets}},
  \code{\link{bdc_filter_mz}}, \code{\link{bdc_filter_rt}},
  \code{\link{bdc_pick_rt_interval}}
}
