% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{bdc_pick_rt_interval}
\alias{bdc_pick_rt_interval}
\title{Title}
\usage{
bdc_pick_rt_interval(mzid)
}
\description{
Title
}
\seealso{
Other bdc functions: \code{\link{bdc_dsets}},
  \code{\link{bdc_filter_mz}}, \code{\link{bdc_filter_rt}},
  \code{\link{bdc_mzid_match}}
}
\keyword{internal}
