% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bdc_rt_tol_rules}
\alias{bdc_rt_tol_rules}
\title{Dataset mapping sets and intervals to tolerances}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 4 columns.}
\usage{
bdc_rt_tol_rules
}
\description{
Used for rt matching. From \url{https://www.dropbox.com/sh/9hc1ai3krsoakbe/AAAkVxDnZp-53FajkXyKkdZga?dl=0&preview=RT_Drift_for_Susan.xlsx}
}
\keyword{datasets}
