% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bdc_set_rules}
\alias{bdc_set_rules}
\title{List mapping datasets to set labels}
\format{An object of class \code{list} of length 5.}
\usage{
bdc_set_rules
}
\description{
Used for rt matching. From \url{https://www.dropbox.com/sh/9hc1ai3krsoakbe/AAAkVxDnZp-53FajkXyKkdZga?dl=0&preview=RT_Drift_for_Susan.xlsx}
}
\keyword{datasets}
